refactor: move pkgdb errors to own module
