+++ crates/flox-rust-sdk/src/models/pkgdb_errors.rs
@@ -1,0 +1,109 @@
+use std::fmt::Display;
+
+use serde::Deserialize;
+use serde_json::Value;
+
+/// A struct representing error messages coming from pkgdb
+#[derive(Debug)]
+pub struct PkgDbError {
+    /// The exit code of pkgdb, can be used to programmatically determine
+    /// the category of error.
+    pub exit_code: u64,
+    /// The generic message for this category of error.
+    pub category_message: String,
+    /// The more contextual message for the specific error that occurred.
+    pub context_message: Option<ContextMsgError>,
+}
+
+impl<'de> Deserialize<'de> for PkgDbError {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        let map = serde_json::Map::<String, Value>::deserialize(deserializer)?;
+        let exit_code = map
+            .get("exit_code")
+            .ok_or_else(|| serde::de::Error::missing_field("exit_code"))?
+            .as_u64()
+            .ok_or_else(|| serde::de::Error::custom("exit code is not an unsigned integer"))?;
+        let category_message = map
+            .get("category_message")
+            .ok_or_else(|| serde::de::Error::missing_field("category_message"))?
+            .as_str()
+            .ok_or_else(|| serde::de::Error::custom("category message was not a string"))
+            .map(|m| m.to_owned())?;
+        let context_message_contents = map
+            .get("context_message")
+            .map(|m| {
+                m.as_str()
+                    .ok_or_else(|| serde::de::Error::custom("context message was not a string"))
+                    .map(|m| m.to_owned())
+            })
+            .transpose()?;
+        let caught_message_contents = map
+            .get("caught_message")
+            .map(|m| {
+                m.as_str()
+                    .ok_or_else(|| serde::de::Error::custom("caught message was not a string"))
+                    .map(|m| m.to_owned())
+            })
+            .transpose()?;
+        let context_message = context_message_contents.map(|m| ContextMsgError {
+            message: m,
+            caught: caught_message_contents.map(|m| CaughtMsgError { message: m }),
+        });
+        Ok(PkgDbError {
+            exit_code,
+            category_message,
+            context_message,
+        })
+    }
+}
+
+impl Display for PkgDbError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.category_message)?;
+        Ok(())
+    }
+}
+
+impl std::error::Error for PkgDbError {
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        self.context_message
+            .as_ref()
+            .map(|s| s as &dyn std::error::Error)
+    }
+}
+
+/// A struct representing the context message from a pkgdb error
+#[derive(Debug, Deserialize)]
+pub struct ContextMsgError {
+    pub message: String,
+    pub caught: Option<CaughtMsgError>,
+}
+
+impl Display for ContextMsgError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.message)
+    }
+}
+
+impl std::error::Error for ContextMsgError {
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        self.caught.as_ref().map(|s| s as &dyn std::error::Error)
+    }
+}
+
+/// A struct representing the caught message from a pkgdb error
+#[derive(Debug, Deserialize)]
+pub struct CaughtMsgError {
+    pub message: String,
+}
+
+impl Display for CaughtMsgError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.message)
+    }
+}
+
+impl std::error::Error for CaughtMsgError {}
