+++ crates/flox-rust-sdk/src/models/search.rs
@@ -8,7 +8,7 @@ use once_cell::sync::Lazy;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
-use super::environment::PkgDbError;
+use super::pkgdb_errors::PkgDbError;
 
 // This is the `PKGDB` path that we actually use.
 // This is set once and prefers the `PKGDB` env variable, but will use
